+++ Plugins/iOS/ApphudUnityAppleSearchAdsWrapper.swift
@@ -1,0 +1,16 @@
+import Foundation
+import ApphudSDK
+import AdServices
+
+@objc final public class ApphudUnityAppleSearchAdsWrapper: NSObject {
+    @MainActor
+    @objc public static func trackAppleSearchAds(callback: @escaping (Bool) -> Void) {
+        if #available(iOS 14.3, *) {
+            Task {
+                if let asaToken = try? AAAttribution.attributionToken() {
+                    Apphud.addAttribution(data: nil, from: .appleAdsAttribution, identifer: asaToken, callback: callback)
+                }
+            }
+        }
+    }
+}