+++ Runtime/SDK/IOS/SDK/ApphudIOSInternal.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Runtime.InteropServices;
 using Apphud.Unity.Common.Utils;
+using Apphud.Unity.Domain;
 
 namespace Apphud.Unity.IOS.SDK
 {
@@ -15,11 +16,11 @@ namespace Apphud.Unity.IOS.SDK
 
         [DllImport("__Internal")]
         private static extern void ApphudUnity_startWithApiKey(string apiKey, bool observerMode, IntPtr callback);
-        internal static void Start(string apiKey, Action<string> callback) => ApphudUnity_startWithApiKey(apiKey, false, callback.ToIntPtr());
+        internal static void Start(string apiKey, Action<string> callback, bool observerMode) => ApphudUnity_startWithApiKey(apiKey, observerMode, callback.ToIntPtr());
 
         [DllImport("__Internal")]
         private static extern void ApphudUnity_startWithApiKeyAndUserId(string apiKey, string userId, bool observerMode, IntPtr callback);
-        internal static void Start(string apiKey, string userId, Action<string> callback) => ApphudUnity_startWithApiKeyAndUserId(apiKey, userId, false, callback.ToIntPtr());
+        internal static void Start(string apiKey, string userId, Action<string> callback, bool observerMode) => ApphudUnity_startWithApiKeyAndUserId(apiKey, userId, observerMode, callback.ToIntPtr());
 
         [DllImport("__Internal")]
         internal static extern void ApphudUnity_logOut();
@@ -84,10 +85,33 @@ namespace Apphud.Unity.IOS.SDK
         internal static extern void ApphudUnity_incrementUserProperty(string key, string byJson);
 
         [DllImport("__Internal")]
-        internal static extern void ApphudUnity_addAttribution(string provider, string dataJson, string identifer);
+        private static extern void ApphudUnity_addAttribution(string provider, string dataJson, string identifer, IntPtr callback);
+        internal static void AddAttribution(ApphudAttributionProvider provider, string dataJson, string identifer, Action<bool> callback) => ApphudUnity_addAttribution(provider.ToString(), dataJson, identifer, callback.ToIntPtr());
+
+        [DllImport("__Internal")]
+        private static extern void ApphudUnity_addFBAttribution(IntPtr callback);
+        internal static void AddFBAttribution(Action<bool> callback) => ApphudUnity_addFBAttribution(callback.ToIntPtr());
+
+        [DllImport("__Internal")]
+        private static extern void ApphudUnity_trackAppleSearchAds(IntPtr callback);
+        internal static void TrackAppleSearchAds(Action<bool> callback) => ApphudUnity_trackAppleSearchAds(callback.ToIntPtr());
+
+        [DllImport("__Internal")]
+        private static extern void ApphudUnity_loadFallbackPaywallsWithCallback(IntPtr callback);
+        internal static void LoadFallbackPaywallsWithCallback(Action<string, string> callback) => ApphudUnity_loadFallbackPaywallsWithCallback(callback.ToIntPtr());
+
+        [DllImport("__Internal")]
+        internal static extern void ApphudUnity_willPurchaseProductFrom(string paywallIdentifier, string placementIdentifier);
 
         [DllImport("__Internal")]
         internal static extern void ApphudUnity_setDeviceIdentifiers(string idfa, string idfv);
+
+        [DllImport("__Internal")]
+        internal static extern void ApphudUnity_setPaywallsCacheTimeout(double value);
+
+        [DllImport("__Internal")]
+        private static extern void ApphudUnity_submitPushNotificationsTokenString(string str, IntPtr callback);
+        internal static void SubmitPushNotificationsTokenString(string str, Action<bool> callback) => ApphudUnity_submitPushNotificationsTokenString(str, callback.ToIntPtr());
 
         [DllImport("__Internal")]
         internal static extern void ApphudUnity_setHeaders();