+++ Runtime/SDK/SDK/ApphudUtils.cs
@@ -1,4 +1,6 @@
-#if UNITY_ANDROID
+#if UNITY_EDITOR
+using Apphud.Unity.Editor;
+#elif UNITY_ANDROID
 using Apphud.Unity.Android.SDK;
 #elif UNITY_IOS
 using Apphud.Unity.IOS.SDK;
@@ -8,7 +10,9 @@ namespace Apphud.Unity.SDK
 {
     public static class ApphudUtils
     {
-#if UNITY_ANDROID
+#if UNITY_EDITOR
+        private static readonly IApphudUtils _utils = new ApphudEditorUtils();
+#elif UNITY_ANDROID
         private static readonly IApphudUtils _utils = new ApphudAndroidUtils();
 #elif UNITY_IOS
         private static readonly IApphudUtils _utils = new ApphudIOSUtils();